requirements.txt
nicegui==1.4.17
slack-sdk==3.27.0
pillow==10.2.0
loguru==0.7.2
python-dotenv==1.0.1
pytest==8.1.1
pytest-asyncio==0.23.5
aiohttp==3.9.3 
